pyproject.toml
[project]
name = "gex-signal-prediction"
version = "0.1.0"
description = "Gamma Exposure (GEX) Signal Prediction System for Nifty & Sensex options"
requires-python = ">=3.10"
dependencies = [
    "upstox-python-sdk>=2.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "pydantic>=2.0",
    "plotly>=5.0",
    "streamlit>=1.30",
    "streamlit-autorefresh>=1.0",
    "requests>=2.31",
    "python-dotenv>=1.0",
    "pyarrow>=14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.pytest.ini_options]
testpaths = ["tests"]
